+++ src/Logging/JsonLogHandler.php
@@ -1,0 +1,6 @@
+<?php
+namespace Mnb\SecurityCore\Logging;
+
+class JsonLogHandler extends FileLogHandler
+{
+}